tmforum-catalyst/consumer-co2-ecs-api/requirements.txt
langchain
sqlalchemy==1.4.47
snowflake-sqlalchemy
urllib3==1.26.6
botocore
requests
#anthropic